// civic-web-app/src/types/analytics.tsx
/** Google Analytics event category. */
export enum GACategory {
  // Events initiated by user.
  USER = "USER",
}

/** Google Analytics event actions that are initiated by user. */
export enum GAUserAction {
  ADD_TO_CALENDAR = "ADD_TO_CALENDAR",
  CLEAR_SELECTED_TAGS = "CLEAR_SELECTED_TAGS",
  CLICK_CIVIC_SOCIAL_LINK = "CLICK_CIVIC_SOCIAL_LINK",
  CLICK_DOWNLOAD_MOBILE_APP = "CLICK_DOWNLOAD_MOBILE_APP",
  CLICK_EVENT_VIEW_MORE_FROM_LIST = "CLICK_EVENT_VIEW_MORE_FROM_LIST",
  CLICK_EVENT_VIEW_MORE_FROM_MAP = "CLICK_EVENT_VIEW_MORE_FROM_MAP",
  CLICK_EXTERNAL_EVENT_LINK = "CLICK_EXTERNAL_EVENT_LINK",
  CLICK_HOME = "CLICK_HOME",
  CLICK_LOGIN = "CLICK_LOGIN",
  CLICK_MAP_MARKER = "CLICK_MAP_MARKER",
  CLICK_READ_LESS_FROM_EVENT_DETAIL = "CLICK_READ_LESS_FROM_EVENT_DETAIL",
  CLICK_READ_MORE_FROM_EVENT_DETAIL = "CLICK_READ_MORE_FROM_EVENT_DETAIL",
  CLICK_SEARCH = "CLICK_SEARCH",
  CLICK_SEE_MORE_EVENT_TYPES = "CLICK_SEE_MORE_EVENT_TYPES",
  CLICK_SEE_MORE_EVENTS_FROM_EVENT_DETAIL =
    "CLICK_SEE_MORE_EVENTS_FROM_EVENT_DETAIL",
  CLOSE_CALENDAR_MODAL = "CLOSE_CALENDAR_MODAL",
  CLOSE_EVENT_CARD = "CLOSE_EVENT_CARD",
  CLOSE_LOGIN_MODAL = "CLOSE_LOGIN_MODAL",
  CLOSE_MOBILE_FILTER_MODAL = "CLOSE_MOBILE_FILTER_MODAL",
  CLOSE_SEARCH_POPPER = "CLOSE_SEARCH_POPPER",
  CLOSE_SHARE_MODAL = "CLOSE_SHARE_MODAL",
  CLOSE_THANK_YOU_MODAL = "CLOSE_THANK_YOU_MODAL",
  COPY_CIVIC_EVENT_LINK = "COPY_CIVIC_EVENT_LINK",
  ENTER_FROM_SEARCH_BAR = "ENTER_FROM_SEARCH_BAR",
  FILTER_BY_DATES = "FILTER_BY_DATES",
  FILTER_BY_TAGS = "FILTER_BY_TAGS",
  FILTER_BY_TYPES = "FILTER_BY_TYPES",
  FILTER_BY_SEARCH_INPUT = "FILTER_BY_SEARCH_INPUT",
  FILTER_BY_VIRTUAL_ONLY = "FILTER_BY_VIRTUAL_ONLY",
  SHARE_FROM_EVENT_DETAIL = "SHARE_FROM_EVENT_DETAIL",
  SHARE_FROM_LIST = "SHARE_FROM_LIST",
  SHARE_FROM_MAP = "SHARE_FROM_MAP",
  SELECT_EVENT_FROM_SEARCH_BAR = "SELECT_EVENT_FROM_SEARCH_BAR",
  SUBSCRIBE_EMAIL = "SUBSCRIBE_EMAIL",
}

/** Google Analytics event labels for various social media. */
export enum GASocialLabel {
  FACEBOOK = "FACEBOOK",
  INSTAGRAM = "INSTAGRAM",
  TWITTER = "TWITTER",
}

/** Google Analytics event labels for different mobile OSes. */
export enum GAMobileOSLabel {
  ANDROID = "ANDROID",
  IOS = "IOS",
}